refactor: extract feeds tool handlers from server.ts into src/mcp/tools/feeds.ts
